feat: Add isDeleted field to user-related models
The user-related models (Users, Positions, Person, Employees, FamilyTypes, FamilyMembers, EmployeeDetails, IncomeTypes, Income, ExpenseTypes, and Expenses) have been updated to include an isDeleted field. This field is set to false by default and can be used to mark a record as deleted without actually removing it from the database. This allows for soft deletion and the ability to restore deleted records if needed.